null: Claim to use 64 bit floating point registers.
The NULL ISA doesn't actually have registers, so this setting doesn't
matter. By making it 64 bits the ISA is more compatible with the other
ISAs.

Reviewed-on: https://gem5-review.googlesource.com/c/13617
Maintainer: Jason Lowe-Power <jason@lowepower.com>

+++ src/arch/null/registers.hh
@@ -47,8 +47,8 @@
 namespace NullISA {
 
 typedef uint64_t IntReg;
-typedef uint32_t FloatRegBits;
-typedef float FloatReg;
+typedef uint64_t FloatRegBits;
+typedef double FloatReg;
 typedef uint8_t CCReg;
 typedef uint64_t MiscReg;
 const RegIndex ZeroReg = 0;
